+++ marketplace/services/product/product_facebook_manage.py
@@ -3,8 +3,11 @@ from typing import List
 from django.contrib.auth import get_user_model
 from django.db import transaction
 
-from marketplace.services.vtex.utils.data_processor import FacebookProductDTO
-from marketplace.wpp_products.models import Product
+from marketplace.wpp_products.models import Product, UploadProduct, Catalog, ProductFeed
+from marketplace.services.vtex.utils.data_processor import (
+    DataProcessor,
+    FacebookProductDTO,
+)
 
 
 User = get_user_model()
@@ -96,4 +99,83 @@ class ProductFacebookManager:
                     f"{len(products_to_create)} products have been created in the integration databases"
                 )
 
+        return True
+
+    def save_csv_product_data(
+        self,
+        products_dto: List[FacebookProductDTO],
+        catalog: Catalog,
+        product_feed: ProductFeed,
+        data_processor: DataProcessor,
+    ):
+        all_success = True
+        for product in products_dto:
+            facebook_product_id = product.id
+            product_csv = data_processor.product_to_csv_line(product)
+
+            try:
+                product, created = UploadProduct.objects.update_or_create(
+                    facebook_product_id=facebook_product_id,
+                    catalog=catalog,
+                    feed=product_feed,
+                    defaults={
+                        "data": product_csv,
+                        "status": "pending",
+                    },
+                )
+                action = "created" if created else "updated"
+                print(
+                    f"Product {action} successfully in the database: Catalog ID {catalog.facebook_catalog_id}, "
+                    f"Feed ID {product_feed.facebook_feed_id}, Product Facebook ID {facebook_product_id}"
+                )
+            except Exception as e:
+                print(f"Failed to save or update product: {str(e)}")
+                all_success = False
+
+        return all_success
+
+    def save_first_csv_product_data(
+        self,
+        products_dto: List[FacebookProductDTO],
+        catalog: Catalog,
+        product_feed: ProductFeed,
+        data_processor: DataProcessor,
+    ):
+        print("Starting the initial data insertion process in UploadProduct")
+
+        # Step 1: Delete existing products for the catalog
+        deleted_count, _ = UploadProduct.objects.filter(catalog=catalog).delete()
+        print(
+            f"Deleted {deleted_count} existing products for catalog {catalog.facebook_catalog_id}"
+        )
+
+        # Step 2: Prepare data for bulk create
+        products_to_create = []
+        batch_size = 30000  # Define the batch size for bulk creation
+
+        for dto in products_dto:
+            product_csv = data_processor.product_to_csv_line(dto)
+            new_product = UploadProduct(
+                facebook_product_id=dto.id,
+                data=product_csv,
+                catalog=catalog,
+                feed=product_feed,
+                status="pending",
+            )
+            products_to_create.append(new_product)
+
+            # Bulk create in batches
+            if len(products_to_create) >= batch_size:
+                UploadProduct.objects.bulk_create(
+                    products_to_create, batch_size=batch_size
+                )
+                print(f"Bulk created {len(products_to_create)} products.")
+                products_to_create.clear()  # Clear the list after bulk creation
+
+        # If remaining products, create
+        if products_to_create:
+            UploadProduct.objects.bulk_create(products_to_create, batch_size=batch_size)
+            print(f"Bulk created {len(products_to_create)} remaining products.")
+
+        print("Initial data insertion process completed successfully.")
         return True