+++ marketplace/services/vtex/utils/data_processor.py
@@ -3,6 +3,7 @@ import pandas as pd
 import io
 import dataclasses
 import threading
+import re
 
 from dataclasses import dataclass
 
@@ -10,6 +11,8 @@ from typing import List
 
 from tqdm import tqdm
 from queue import Queue
+
+from marketplace.clients.exceptions import CustomAPIException
 
 
 @dataclass
@@ -37,6 +40,22 @@ class DataProcessor:
     def __init__(self):
         self.max_workers = 100
         self.progress_lock = threading.Lock()
+
+    @staticmethod
+    def clean_text(text: str) -> str:
+        """Cleans up text by removing HTML tags, replacing quotes with empty space,
+        replacing commas with semicolons, and normalizing whitespace."""
+        # Remove HTML tags
+        text = re.sub(r"<[^>]*>", "", text)
+        # Replace double and single quotes with empty space
+        text = text.replace('"', "").replace("'", " ")
+        # Replace commas with semicolons
+        text = text.replace(",", ";")
+        # Normalize new lines and carriage returns to space and remove excessive whitespace
+        text = re.sub(r"\s+", " ", text.strip())
+        # Remove bullet points
+        text = text.replace("•", "")
+        return text
 
     @staticmethod
     def extract_fields(
@@ -67,9 +86,12 @@ class DataProcessor:
             else product_details["SkuName"]
         )
         title = product_details["SkuName"].title()
-        # Limit title and description to 200 characters for Facebook rules
-        description = description[:200]
-        title = title[:200]
+        # Applies the .title() before clearing the text
+        title = title[:200].title()
+        description = description[:9999].title()
+        # Clean title and description
+        title = DataProcessor.clean_text(title)
+        description = DataProcessor.clean_text(description)
 
         availability = (
             "in stock" if availability_details["is_available"] else "out of stock"
@@ -78,8 +100,8 @@ class DataProcessor:
 
         return FacebookProductDTO(
             id=sku_id,
-            title=title.title(),
-            description=description.title(),
+            title=title,
+            description=description,
             availability=availability,
             status=status,
             condition="new",
@@ -162,15 +184,30 @@ class DataProcessor:
 
     def process_single_sku(self, sku_id):
         facebook_products = []
-        product_details = self.service.get_product_details(sku_id, self.domain)
+
+        try:
+            product_details = self.service.get_product_details(sku_id, self.domain)
+        except CustomAPIException as e:
+            if e.status_code == 404:
+                print(f"SKU {sku_id} not found. Skipping...")
+            return []
+
         is_active = product_details.get("IsActive")
         if not is_active and not self.update_product:
             return facebook_products
 
         for seller_id in self.active_sellers:
-            availability_details = self.service.simulate_cart_for_seller(
-                sku_id, seller_id, self.domain
-            )
+            try:
+                availability_details = self.service.simulate_cart_for_seller(
+                    sku_id, seller_id, self.domain
+                )
+            except CustomAPIException as e:
+                if e.status_code == 500:
+                    print(
+                        f"An error {e.status_code} occurred when simulating cart. SKU {sku_id}, Seller {seller_id}."
+                        "Skipping..."
+                    )
+                continue
             if (
                 self.update_product is False
                 and not availability_details["is_available"]
@@ -206,6 +243,14 @@ class DataProcessor:
         buffer.seek(0)
         print("CSV file successfully generated in memory")
         return buffer
+
+    @staticmethod
+    def product_to_csv_line(product: FacebookProductDTO) -> str:
+        product_dict = dataclasses.asdict(product)
+        df = pd.DataFrame([product_dict])
+        df = df.drop(columns=["product_details"])
+        csv_line = df.to_csv(index=False, header=False, encoding="utf-8").strip()
+        return csv_line
 
     @staticmethod
     def clear_csv_buffer(buffer: io.BytesIO):