+++ CHANGELOG.md
@@ -1,3 +1,58 @@
+v3.7.4
+----------
+* Hotfix: Create alert to notify upload errors
+
+v3.7.3
+----------
+* Hotfix: Remove quotation marks and HTML tags from the title and description of vtex products
+
+v3.7.2
+----------
+* Check the view if it can synchronize and add ignore results to tasks
+
+v3.7.1
+----------
+* Hotfix: Set CELERY_RESULT_BACKEND with None
+
+v3.7.0
+----------
+* Processes vtex update webhooks asynchronously and saves them on DB
+* Create task to upload products from db to fba-meta
+* Saves uploads sent to Facebook and creates a task to clear logs 
+* Fix: limits requests on vtex per minute
+* Fix: list paginated sellers
+* Remove url redirection from view
+
+v3.6.4
+----------
+* Temporarily enable redirection of vtex updates
+
+v3.6.3
+----------
+* Ignore 404 error when fetching vtex product details
+* Round up prices for weight-calculated products
+* First sync vtex products for specific sellers
+* Vtex rules for colombia
+* Add method to list only active skus IDs
+
+v3.6.2
+----------
+* Remove vtex webhook log
+
+v3.6.1
+----------
+* Add vtex webhook log and add retry in facebook methods
+
+v3.6.0
+----------
+* Permission update consumer
+* Check if there is a feed upload in progress before uploading a new
+
+v3.5.0
+----------
+* Sends webhook and template data to flows
+* Fix test_update_template_translation_error test case
+
 v3.4.6
 ----------
 * Updates template per app and ignores apps with errors in waba